🚚 Move remaining entities from `input/mod.rs` into `input/core.rs`
This way `mod.rs` finally only has module declarations.
