+++ examples/evaluate_expression_string.rs
@@ -1,9 +1,9 @@
+use expression::{Expression, Operator, evaluate};
 use std::io;
 use yapcol::error::Error;
-use yapcol::input::Input;
+use yapcol::input::core::Input;
 use yapcol::{StringParser, attempt, between, chain_left, chain_right, is, many0, option, satisfy};
 mod expression;
-use expression::{Expression, Operator, evaluate};
 
 trait StringExpressionParser: StringParser<Expression> {}
 
@@ -98,7 +98,7 @@ fn main() {
 		match stdin.read_line(input) {
 			Ok(_) if input.len() == 2 && input.starts_with('q') => break,
 			Ok(_) => {
-				let mut input = Input::new(input.chars());
+				let mut input = Input::new_from_chars(input.chars());
 				match parse_expression()(&mut input) {
 					Ok(e) => println!("Success: {:?}", evaluate(e)),
 					Err(e) => println!("Failed to parse expression: {e}"),
@@ -122,7 +122,7 @@ mod parsing_tests {
 
 	#[test]
 	fn number() {
-		let mut input = Input::new("123".chars());
+		let mut input = Input::new_from_chars("123".chars());
 		let parser = parse_expression();
 		assert_eq!(parser(&mut input), Ok(Expression::Number(123)));
 	}
@@ -132,7 +132,7 @@ mod parsing_tests {
 		let number1 = 123;
 		let number2 = 456;
 		let tokens = format!("{number1}+{number2}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		assert_eq!(
@@ -147,7 +147,7 @@ mod parsing_tests {
 		let number1 = 123;
 		let number2 = 456;
 		let tokens = format!("{number1}-{number2}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		assert_eq!(
@@ -162,7 +162,7 @@ mod parsing_tests {
 		let number1 = 123;
 		let number2 = 456;
 		let tokens = format!("{number1}*{number2}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		assert_eq!(
@@ -177,7 +177,7 @@ mod parsing_tests {
 		let number1 = 123;
 		let number2 = 456;
 		let tokens = format!("{number1}/{number2}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		assert_eq!(
@@ -193,7 +193,7 @@ mod parsing_tests {
 		let number2 = 456;
 		let number3 = 789;
 		let tokens = format!("{number1}+{number2}+{number3}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		// Addition is left-associative.
@@ -213,7 +213,7 @@ mod parsing_tests {
 		let number2 = 456;
 		let number3 = 789;
 		let tokens = format!("{number1}+{number2}-{number3}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		let expression1 = build_operation(number1, Operator::Addition, number2);
@@ -232,7 +232,7 @@ mod parsing_tests {
 		let number2 = 456;
 		let number3 = 789;
 		let tokens = format!("{number1}+{number2}*{number3}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		let expression1 = build_operation(number2, Operator::Multiplication, number3);
@@ -250,7 +250,7 @@ mod parsing_tests {
 		let number1 = 12;
 		let number2 = 3;
 		let tokens = format!("{number1}^{number2}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		assert_eq!(
@@ -266,7 +266,7 @@ mod parsing_tests {
 		let number2 = 3;
 		let number3 = 4;
 		let tokens = format!("{number1}^{number2}^{number3}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let output = parser(&mut input).unwrap();
 		let expression1 = build_operation(number2, Operator::Exponentiation, number3);
@@ -282,14 +282,14 @@ mod parsing_tests {
 
 	#[test]
 	fn parenthesis_number() {
-		let mut input = Input::new("(123)".chars());
+		let mut input = Input::new_from_chars("(123)".chars());
 		let parser = parse_expression();
 		assert_eq!(parser(&mut input), Ok(Expression::Number(123)));
 	}
 
 	#[test]
 	fn double_parenthesis_number() {
-		let mut input = Input::new("((123))".chars());
+		let mut input = Input::new_from_chars("((123))".chars());
 		let parser = parse_expression();
 		assert_eq!(parser(&mut input), Ok(Expression::Number(123)));
 	}
@@ -300,7 +300,7 @@ mod parsing_tests {
 		let number2 = 456;
 		let number3 = 789;
 		let tokens = format!("({number1}+{number2})*{number3}");
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let parser = parse_expression();
 		let expression1 = Expression::Operation(
 			Box::new(Expression::Number(number1)),
@@ -323,7 +323,7 @@ mod evaluation_tests {
 	use yapcol::end_of_input;
 
 	fn parse_and_evaluate(tokens: &str) -> i32 {
-		let mut input = Input::new(tokens.chars());
+		let mut input = Input::new_from_chars(tokens.chars());
 		let output = parse_expression()(&mut input);
 		assert!(end_of_input()(&mut input).is_ok());
 		evaluate(output.unwrap())
