+++ examples/evaluate_expression_token.rs
@@ -1,10 +1,10 @@
+use expression::{Expression, Operator, evaluate};
 use std::io;
 use yapcol::error::Error;
+use yapcol::input::core::{Input, InputToken};
 use yapcol::input::position::Position;
-use yapcol::input::{Input, InputToken};
 use yapcol::{Parser, attempt, between, chain_left, chain_right, is, option, satisfy};
 mod expression;
-use expression::{Expression, Operator, evaluate};
 
 #[derive(Debug, PartialEq, Clone)]
 enum Token {
