+++ src/lib.rs
@@ -22,10 +22,10 @@
 //! # Quick Start
 //!
 //! ```
-//! use yapcol::input::Input;
+//! use yapcol::input::core::{Input};
 //! use yapcol::{is, many0};
 //!
-//! let mut input = Input::new("aaab".chars());
+//! let mut input = Input::new_from_chars("aaab".chars());
 //!
 //! // Combine `is` and `many0` to parse multiple 'a's
 //! let is_a = is('a');
@@ -51,8 +51,8 @@
 //! more information.
 
 use crate::error::Error;
+use crate::input::core::{Input, InputToken};
 use crate::input::string::{CharToken, StringInput};
-use crate::input::{Input, InputToken};
 
 pub mod error;
 pub mod input;
@@ -79,7 +79,7 @@ mod tests;
 ///
 /// ```
 /// use std::str::Chars;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 /// use yapcol::input::string::StringInput;
 /// use yapcol::error::Error;
 /// use yapcol::is;
@@ -89,17 +89,17 @@ mod tests;
 ///    is('A')(input)
 /// }
 ///
-/// let mut input = Input::new("Abc".chars());
+/// let mut input = Input::new_from_chars("Abc".chars());
 /// assert_eq!(my_uppercase_parser(&mut input), Ok('A'));
 /// ```
 ///
 /// Most of the time, you will use the built-in combinators which return `impl Parser`:
 ///
 /// ```
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 /// use yapcol::is;
 ///
-/// let mut input = Input::new("Abc".chars());
+/// let mut input = Input::new_from_chars("Abc".chars());
 /// let mut parser = is('A');
 /// assert_eq!(parser(&mut input), Ok('A'));
 /// ```
@@ -133,15 +133,15 @@ impl<O, X> StringParser<O> for X where X: Fn(&mut StringInput) -> Result<O, Erro
 ///
 /// ```
 /// use yapcol::{is, any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// let tokens: Vec<char> = vec!['h', 'e', 'l', 'l', 'o'];
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// let parser = is('h');
 /// assert!(parser(&mut input).is_ok());
 ///
 /// let mut wrong: Vec<char> = vec!['w', 'o', 'r', 'l', 'd'];
-/// let mut input = Input::new(wrong);
+/// let mut input = Input::new_from_chars(wrong);
 /// assert!(parser(&mut input).is_err());
 /// assert_eq!(any()(&mut input), Ok('w')); // Input was not consumed.
 /// ```
@@ -174,10 +174,10 @@ where
 /// ```
 /// use yapcol::{satisfy, any};
 /// use yapcol::error::Error;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// let tokens: Vec<char> = vec!['3', 'a', 'b'];
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// let parser = satisfy(|c: &char| {
 ///     if c.is_ascii_digit() { Some(*c) } else { None }
 /// });
@@ -185,7 +185,7 @@ where
 /// assert_eq!(any()(&mut input), Ok('a')); // Token was consumed.
 ///
 /// let tokens: Vec<char> = vec!['a', 'b', 'c'];
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// assert!(parser(&mut input).is_err());
 /// assert_eq!(any()(&mut input), Ok('a')); // Input was not consumed.
 /// ```
@@ -219,14 +219,14 @@ where
 /// ```
 /// use yapcol::{end_of_input, any};
 /// use yapcol::error::Error;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// let tokens: Vec<char> = vec![];
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// assert!(end_of_input()(&mut input).is_ok());
 ///
 /// let tokens: Vec<char> = vec!['a', 'b'];
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// assert!(end_of_input()(&mut input).is_err());
 /// assert_eq!(any()(&mut input), Ok('a')); // Input was not consumed.
 /// ```
@@ -259,20 +259,20 @@ where
 ///
 /// ```
 /// use yapcol::{option, is, end_of_input, any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// // parser1 succeeds: returns its result.
-/// let mut input = Input::new("ab".chars());
+/// let mut input = Input::new_from_chars("ab".chars());
 /// let output = option(&is('a'), &is('b'))(&mut input).unwrap();
 /// assert_eq!(output, 'a');
 ///
 /// // parser1 fails without consuming input: parser2 is tried.
-/// let mut input = Input::new("b".chars());
+/// let mut input = Input::new_from_chars("b".chars());
 /// let output = option(&is('a'), &is('b'))(&mut input).unwrap();
 /// assert_eq!(output, 'b');
 ///
 /// // Both parsers fail: an error is returned and input is not consumed.
-/// let mut input = Input::new("c".chars());
+/// let mut input = Input::new_from_chars("c".chars());
 /// assert!(option(&is('a'), &is('b'))(&mut input).is_err());
 /// assert_eq!(any()(&mut input), Ok('c'));
 /// ```
@@ -306,14 +306,14 @@ where
 ///
 /// ```
 /// use yapcol::{maybe, is, any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
-/// let mut input = Input::new("hello".chars());
+/// let mut input = Input::new_from_chars("hello".chars());
 /// let ph = is('h');
 /// let parser = maybe(&ph);
 /// assert_eq!(parser(&mut input).unwrap(), Some('h'));
 ///
-/// let mut input = Input::new("world".chars());
+/// let mut input = Input::new_from_chars("world".chars());
 /// assert_eq!(parser(&mut input).unwrap(), None);
 /// assert_eq!(any()(&mut input), Ok('w')); // Input was not consumed.
 /// ```
@@ -357,19 +357,19 @@ where
 ///
 /// ```
 /// use yapcol::{many0, is};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// // Matches multiple elements
 /// let parser = is('1');
-/// let mut input = Input::new("112".chars());
+/// let mut input = Input::new_from_chars("112".chars());
 /// assert_eq!(many0(&parser)(&mut input), Ok("11".chars().collect()));
 ///
 /// // Returns an empty vector when no matches are found (never fails)
-/// let mut input = Input::new("23".chars());
+/// let mut input = Input::new_from_chars("23".chars());
 /// assert_eq!(many0(&parser)(&mut input), Ok(vec![]));
 ///
 /// // Returns an empty vector on empty input (never fails)
-/// let mut input = Input::new("".chars());
+/// let mut input = Input::new_from_chars("".chars());
 /// assert_eq!(many0(&parser)(&mut input), Ok(vec![]));
 /// ```
 pub fn many0<P, IT, O>(parser: &P) -> impl Parser<IT, Vec<O>>
@@ -394,19 +394,19 @@ where
 ///
 /// ```
 /// use yapcol::{many1, is};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// // Matches multiple elements
 /// let parser = is('1');
-/// let mut input = Input::new("112".chars());
+/// let mut input = Input::new_from_chars("112".chars());
 /// assert_eq!(many1(&parser)(&mut input), Ok("11".chars().collect()));
 ///
 /// // Fails when no matches are found
-/// let mut input = Input::new("23".chars());
+/// let mut input = Input::new_from_chars("23".chars());
 /// assert!(many1(&parser)(&mut input).is_err());
 ///
 /// // Fails on empty input
-/// let mut input = Input::new("".chars());
+/// let mut input = Input::new_from_chars("".chars());
 /// assert!(many1(&parser)(&mut input).is_err());
 /// ```
 pub fn many1<P, IT, O>(parser: &P) -> impl Parser<IT, Vec<O>>
@@ -437,21 +437,21 @@ where
 ///
 /// ```
 /// use yapcol::{choice, is};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// // Returns the result of the first matching parser
 /// let p1 = is('1');
 /// let p2 = is('2');
 /// let parsers = vec![p1, p2];
-/// let mut input = Input::new("23".chars());
+/// let mut input = Input::new_from_chars("23".chars());
 /// assert_eq!(choice(&parsers)(&mut input), Ok('2'));
 ///
 /// // Fails when no parser matches
-/// let mut input = Input::new("34".chars());
+/// let mut input = Input::new_from_chars("34".chars());
 /// assert!(choice(&parsers)(&mut input).is_err());
 ///
 /// // Fails on empty input
-/// let mut input = Input::new("".chars());
+/// let mut input = Input::new_from_chars("".chars());
 /// assert!(choice(&parsers)(&mut input).is_err());
 /// ```
 pub fn choice<'a, P, IT, O, PI>(parsers: &'a PI) -> impl Parser<IT, O>
@@ -483,24 +483,24 @@ where
 ///
 /// ```
 /// use yapcol::{count, is, any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
 /// // Succeeds when the parser matches exactly `count` times
 /// let parser = is('1');
-/// let mut input = Input::new("1112".chars());
+/// let mut input = Input::new_from_chars("1112".chars());
 /// assert_eq!(count(&parser, 3)(&mut input), Ok("111".chars().collect()));
 /// assert_eq!(any()(&mut input), Ok('2')); // Remaining input after consuming 3 tokens.
 ///
 /// // Fails when there are not enough matching tokens
-/// let mut input = Input::new("123".chars());
+/// let mut input = Input::new_from_chars("123".chars());
 /// assert!(count(&parser, 3)(&mut input).is_err());
 ///
 /// // Succeeds with count = 0, returning an empty vector
-/// let mut input = Input::new("123".chars());
+/// let mut input = Input::new_from_chars("123".chars());
 /// assert_eq!(count(&parser, 0)(&mut input), Ok(vec![]));
 ///
 /// // Fails on empty input when count > 0
-/// let mut input = Input::new("".chars());
+/// let mut input = Input::new_from_chars("".chars());
 /// assert!(count(&parser, 1)(&mut input).is_err());
 /// ```
 pub fn count<P, IT, O>(parser: &P, count: usize) -> impl Parser<IT, Vec<O>>
@@ -533,12 +533,12 @@ where
 ///
 /// ```
 /// use yapcol::{look_ahead, is, end_of_input, any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 /// use yapcol::input::position::Position;
 /// use yapcol::error::Error;
 ///
 /// // Succeeds without consuming input.
-/// let mut input = Input::new("123".chars());
+/// let mut input = Input::new_from_chars("123".chars());
 /// let parser1 = is('1');
 /// assert_eq!(look_ahead(&parser1)(&mut input), Ok('1'));
 /// assert_eq!(any()(&mut input), Ok('1')); // Input was not consumed.
@@ -546,12 +546,12 @@ where
 /// assert_eq!(any()(&mut input), Ok('3')); // Input was not consumed.
 ///
 /// // Fails without consuming input.
-/// let mut input = Input::new("23".chars());
+/// let mut input = Input::new_from_chars("23".chars());
 /// assert!(look_ahead(&parser1)(&mut input).is_err());
 /// assert_eq!(any()(&mut input), Ok('2')); // Input was not consumed.
 ///
 /// // Fails consuming input if the parser consumes.
-/// let mut input = Input::new("13".chars());
+/// let mut input = Input::new_from_chars("13".chars());
 /// let consuming_parser = |input: &mut Input<_>| {
 ///     let o1 = parser1(input)?;
 ///     let o2 = parser1(input)?;
@@ -562,7 +562,7 @@ where
 /// assert_eq!(any()(&mut input), Ok('3')); // Input was consumed.
 ///
 /// // Fails on empty input.
-/// let mut input = Input::new("".chars());
+/// let mut input = Input::new_from_chars("".chars());
 /// assert!(look_ahead(&parser1)(&mut input).is_err());
 /// ```
 pub fn look_ahead<P, IT, O>(parser: &P) -> impl Parser<IT, O>
@@ -605,24 +605,24 @@ where
 ///
 /// ```
 /// use yapcol::{attempt, is, end_of_input, any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 /// use yapcol::input::position::Position;
 /// use yapcol::error::Error;
 ///
 /// // Succeeds consuming input.
-/// let mut input = Input::new("123".chars());
+/// let mut input = Input::new_from_chars("123".chars());
 /// let parser1 = is('1');
 /// assert_eq!(attempt(&parser1)(&mut input), Ok('1'));
 /// assert_eq!(any()(&mut input), Ok('2')); // Input was consumed.
 ///
 /// // Fails without consuming input.
-/// let mut input = Input::new("23".chars());
+/// let mut input = Input::new_from_chars("23".chars());
 /// assert!(attempt(&parser1)(&mut input).is_err());
 /// assert_eq!(any()(&mut input), Ok('2')); // Input was not consumed.
 /// assert_eq!(any()(&mut input), Ok('3')); // Input was not consumed.
 ///
 /// // Fails without consuming input if the parser consumes.
-/// let mut input = Input::new("13".chars());
+/// let mut input = Input::new_from_chars("13".chars());
 /// let consuming_parser = |input: &mut Input<_>| {
 ///     let o1 = parser1(input)?;
 ///     let o2 = parser1(input)?;
@@ -633,7 +633,7 @@ where
 /// assert_eq!(any()(&mut input), Ok('1')); // Input was not consumed.
 ///
 /// // Fails on empty input
-/// let mut input = Input::new("".chars());
+/// let mut input = Input::new_from_chars("".chars());
 /// assert!(attempt(&parser1)(&mut input).is_err());
 /// ```
 pub fn attempt<P, IT, O>(parser: &P) -> impl Parser<IT, O>
@@ -661,9 +661,9 @@ where
 ///
 /// ```
 /// use yapcol::{is, between};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::{Input};
 ///
-/// let mut input = Input::new("121".chars());
+/// let mut input = Input::new_from_chars("121".chars());
 /// let parser1 = is('1');
 /// let parser2 = is('2');
 /// let output = between(&parser1, &parser2, &parser1)(&mut input);
@@ -690,10 +690,10 @@ where
 ///
 /// ```
 /// use yapcol::{any};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 ///
 /// // An example input iterator
-/// let mut input = Input::new("123".chars());
+/// let mut input = Input::new_from_chars("123".chars());
 /// let output = any()(&mut input);
 /// assert_eq!(output, Ok('1'));
 /// ```
@@ -736,11 +736,11 @@ where
 ///
 /// ```
 /// use yapcol::{is, separated_by0};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 ///
 /// let parser1 = is('1');
 /// let parser2 = is('2');
-/// let mut input = Input::new("121".chars());
+/// let mut input = Input::new_from_chars("121".chars());
 /// let parser_separated_by0 = separated_by0(&parser1, &parser2);
 /// let output = parser_separated_by0(&mut input);
 /// assert_eq!(output, Ok("11".chars().collect()));
@@ -772,11 +772,11 @@ where
 ///
 /// ```
 /// use yapcol::{is, separated_by1};
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 ///
 /// let parser1 = is('1');
 /// let parser2 = is('2');
-/// let mut input = Input::new("121".chars());
+/// let mut input = Input::new_from_chars("121".chars());
 /// let parser_separated_by1 = separated_by1(&parser1, &parser2);
 /// let output = parser_separated_by1(&mut input);
 /// assert_eq!(output, Ok("11".chars().collect()));
@@ -833,7 +833,7 @@ where
 /// // Implements evaluation of the subtraction ('-') operator as left-associative.
 /// use yapcol::{satisfy, chain_left};
 /// use yapcol::error::Error;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 ///
 /// let operand = satisfy(|c: &char| c.to_digit(10));
 ///
@@ -843,7 +843,7 @@ where
 /// });
 ///
 /// let tokens: Vec<_> = "3-1-1".chars().collect();
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// let parser = chain_left(&operand, &operator);
 /// let output = parser(&mut input);
 /// assert_eq!(output, Ok(1)); // (3 - 1) - 1 = 1, not 3 - (1 - 1) = 3
@@ -879,7 +879,7 @@ where
 /// // Implements evaluation of the subtraction ('-') operator as left-associative.
 /// use yapcol::{satisfy, chain_right};
 /// use yapcol::error::Error;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 ///
 /// let operand = satisfy(|c: &char| c.to_digit(10));
 ///
@@ -889,7 +889,7 @@ where
 /// });
 ///
 /// let tokens: Vec<_> = "3-1-1".chars().collect();
-/// let mut input = Input::new(tokens);
+/// let mut input = Input::new_from_chars(tokens);
 /// let parser = chain_right(&operand, &operator);
 /// let output = parser(&mut input);
 /// assert_eq!(output, Ok(3)); // 3 - (1 - 1) = 3, not (3 - 1) - 1 = 1
@@ -925,16 +925,16 @@ where
 /// ```
 /// use yapcol::{is, not_followed_by};
 /// use yapcol::error::Error;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 /// use yapcol::input::position::Position;
 ///
 /// let parser = is('j');
-/// let mut input = Input::new("hello".chars());
+/// let mut input = Input::new_from_chars("hello".chars());
 /// let not_followed_parser = not_followed_by(&parser);
 /// let output = not_followed_parser(&mut input);
 /// assert_eq!(output, Ok(()));
 ///
-/// let mut input = Input::new("jello".chars());
+/// let mut input = Input::new_from_chars("jello".chars());
 /// let output = not_followed_parser(&mut input);
 /// assert_eq!(output, Err(Error::UnexpectedToken(Position::new(1,1))));
 ///
@@ -968,7 +968,7 @@ where
 /// ```
 /// use yapcol::{any, is, many_until};
 /// use yapcol::error::Error;
-/// use yapcol::input::Input;
+/// use yapcol::input::core::Input;
 ///
 /// let comments_parser = |input: &mut Input<_>| {
 ///     let open = is('#');
@@ -977,7 +977,7 @@ where
 ///     open(input)?;
 ///     many_until(&any, &close)(input)
 /// };
-/// let mut input = Input::new("#this is a comment$".chars());
+/// let mut input = Input::new_from_chars("#this is a comment$".chars());
 /// let output = comments_parser(&mut input);
 /// assert_eq!(output, Ok("this is a comment".chars().collect()));
 /// ```
