+++ src/tests/any.rs
@@ -2,14 +2,14 @@ use crate::*;
 
 #[test]
 fn empty() {
-	let mut input = Input::new("".chars());
+	let mut input = Input::new_from_chars("".chars());
 	let output = any()(&mut input);
 	assert_eq!(output, Err(Error::EndOfInput));
 }
 
 #[test]
 fn success() {
-	let mut input = Input::new("abc".chars());
+	let mut input = Input::new_from_chars("abc".chars());
 	let output = any()(&mut input);
 	assert_eq!(output, Ok('a'));
 	// Using it twice returns the second token.
