+++ src/tests/chain_right.rs
@@ -16,28 +16,28 @@ fn parse_evaluate_right_subtraction() -> impl Parser<CharToken, i32> {
 
 #[test]
 fn empty() {
-	let mut input = Input::new("".chars());
+	let mut input = Input::new_from_chars("".chars());
 	let output = parse_evaluate_right_subtraction()(&mut input);
 	assert_eq!(output, Err(Error::EndOfInput));
 }
 
 #[test]
 fn one_operand() {
-	let mut input = Input::new("1".chars());
+	let mut input = Input::new_from_chars("1".chars());
 	let output = parse_evaluate_right_subtraction()(&mut input);
 	assert_eq!(output, Ok(1));
 }
 
 #[test]
 fn two_operands() {
-	let mut input = Input::new("9-7".chars());
+	let mut input = Input::new_from_chars("9-7".chars());
 	let output = parse_evaluate_right_subtraction()(&mut input);
 	assert_eq!(output, Ok(2));
 }
 
 #[test]
 fn tree_operands() {
-	let mut input = Input::new("3-1-2".chars()); // 3 - (1 - 2) = 4, not (3 - 1) - 2 = 0
+	let mut input = Input::new_from_chars("3-1-2".chars()); // 3 - (1 - 2) = 4, not (3 - 1) - 2 = 0
 	let output = parse_evaluate_right_subtraction()(&mut input);
 	assert_eq!(output, Ok(4));
 }
