+++ src/tests/choice.rs
@@ -10,22 +10,22 @@ fn success() {
 		vec![Box::new(parser1), Box::new(parser2), Box::new(parser3)];
 	let parser_choice = choice(&parsers);
 	// 1, success.
-	let mut input = Input::new("h".chars());
+	let mut input = Input::new_from_chars("h".chars());
 	let output = parser_choice(&mut input).unwrap();
 	assert_eq!(output, 'h');
 	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
 	// 2, success.
-	let mut input = Input::new("e".chars());
+	let mut input = Input::new_from_chars("e".chars());
 	let output = parser_choice(&mut input).unwrap();
 	assert_eq!(output, 'e');
 	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
 	// 3, success.
-	let mut input = Input::new("l".chars());
+	let mut input = Input::new_from_chars("l".chars());
 	let output = parser_choice(&mut input).unwrap();
 	assert_eq!(output, 'l');
 	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
 	// 4, fail.
-	let mut input = Input::new("u".chars());
+	let mut input = Input::new_from_chars("u".chars());
 	assert_eq!(
 		parser_choice(&mut input),
 		Err(Error::UnexpectedToken(Position::new(1, 1)))
@@ -42,7 +42,7 @@ fn fail() {
 		vec![Box::new(parser1), Box::new(parser2), Box::new(parser3)];
 	let parser_choice = choice(&parsers);
 	// 1, success.
-	let mut input = Input::new("x".chars());
+	let mut input = Input::new_from_chars("x".chars());
 	let output = parser_choice(&mut input);
 	assert_eq!(output, Err(Error::UnexpectedToken(Position::new(1, 1))));
 	assert!(end_of_input()(&mut input).is_err()); // Ensure that the input was NOT consumed.
