+++ src/tests/end_of_input.rs
@@ -2,12 +2,12 @@ use crate::*;
 
 #[test]
 fn success() {
-	let mut input = Input::new("".chars());
+	let mut input = Input::new_from_chars("".chars());
 	assert!(end_of_input()(&mut input).is_ok());
 }
 
 #[test]
 fn fail() {
-	let mut input = Input::new("hello".chars());
+	let mut input = Input::new_from_chars("hello".chars());
 	assert!(end_of_input()(&mut input).is_err());
 }
