+++ src/tests/is.rs
@@ -4,7 +4,7 @@ use input::position::Position;
 #[test]
 fn success() {
 	let parser = is('h');
-	let mut input = Input::new("h".chars());
+	let mut input = Input::new_from_chars("h".chars());
 	assert_eq!(parser(&mut input), Ok('h'));
 	assert!(end_of_input()(&mut input).is_ok()); // Ensure that the input was consumed.
 }
@@ -12,7 +12,7 @@ fn success() {
 #[test]
 fn fail() {
 	let parser = is('j');
-	let mut input = Input::new("h".chars());
+	let mut input = Input::new_from_chars("h".chars());
 	assert_eq!(
 		parser(&mut input),
 		Err(Error::UnexpectedToken(Position::new(1, 1)))
