+++ src/tests/look_ahead.rs
@@ -4,7 +4,7 @@ use input::position::Position;
 #[test]
 fn empty() {
 	let parser = is('h');
-	let mut input = Input::new("".chars());
+	let mut input = Input::new_from_chars("".chars());
 	let parse_look_ahead = look_ahead(&parser)(&mut input);
 	assert_eq!(parse_look_ahead, Err(Error::EndOfInput));
 }
@@ -12,7 +12,7 @@ fn empty() {
 #[test]
 fn success_does_not_consume() {
 	let parser = is('h');
-	let mut input = Input::new("hello".chars());
+	let mut input = Input::new_from_chars("hello".chars());
 	let output = look_ahead(&parser)(&mut input);
 	assert_eq!(output, Ok('h'));
 	// After look_ahead, input should still start with hello
@@ -27,7 +27,7 @@ fn success_does_not_consume() {
 #[test]
 fn non_consuming_fail_does_not_consume() {
 	let parser = is('h');
-	let mut input = Input::new("j".chars());
+	let mut input = Input::new_from_chars("j".chars());
 	let output = look_ahead(&parser)(&mut input);
 	assert_eq!(output, Err(Error::UnexpectedToken(Position::new(1, 1))));
 	// Input should still be intact.
@@ -37,7 +37,7 @@ fn non_consuming_fail_does_not_consume() {
 
 #[test]
 fn consuming_fail_consumes() {
-	let mut input = Input::new("he".chars());
+	let mut input = Input::new_from_chars("he".chars());
 	let parser = |input: &mut Input<_>| {
 		let output1 = is('h')(input)?; // Success, therefore it consumed.
 		let output2 = is('a')(input)?; // Failed, so the whole parser fails consuming.
@@ -53,7 +53,7 @@ fn consuming_fail_consumes() {
 #[test]
 fn parse_does_not_consume_on_failure() {
 	let parser = is('h');
-	let mut input = Input::new("jello".chars());
+	let mut input = Input::new_from_chars("jello".chars());
 	let result = look_ahead(&parser)(&mut input);
 	assert_eq!(result, Err(Error::UnexpectedToken(Position::new(1, 1))));
 	// Input should still be intact
@@ -68,7 +68,7 @@ fn parse_does_not_consume_on_failure() {
 #[test]
 fn parse_look_ahead_twice() {
 	let parser = is('h');
-	let mut input = Input::new("h".chars());
+	let mut input = Input::new_from_chars("h".chars());
 	let first = look_ahead(&parser)(&mut input);
 	let second = look_ahead(&parser)(&mut input);
 	assert_eq!(first, Ok('h'));
