+++ src/tests/many0.rs
@@ -3,7 +3,7 @@ use crate::*;
 #[test]
 fn empty() {
 	let parser = is('h');
-	let mut input = Input::new("".chars());
+	let mut input = Input::new_from_chars("".chars());
 	let parser_many0 = many0(&parser);
 	let output = parser_many0(&mut input).unwrap();
 	assert_eq!(output.len(), 0);
@@ -14,7 +14,7 @@ fn no_match_not_empty() {
 	let token_count = 100;
 	let parser = is('h');
 	let tokens = std::iter::repeat_n('j', token_count).collect::<Vec<_>>();
-	let mut input = Input::new(tokens);
+	let mut input = Input::new_from_chars(tokens);
 	let parser_many0 = many0(&parser);
 	let output = parser_many0(&mut input).unwrap();
 	assert_eq!(output.len(), 0);
@@ -27,7 +27,7 @@ fn match_not_empty() {
 	let token_count = 100;
 	let parser = is('h');
 	let tokens = std::iter::repeat_n('h', token_count).collect::<Vec<_>>();
-	let mut input = Input::new(tokens);
+	let mut input = Input::new_from_chars(tokens);
 	let parser_many0 = many0(&parser);
 	let output = parser_many0(&mut input).unwrap();
 	assert_eq!(output.len(), token_count);
