+++ src/tests/not_followed_by.rs
@@ -4,7 +4,7 @@ use input::position::Position;
 #[test]
 fn empty() {
 	let parser = is('h');
-	let mut input = Input::new("".chars());
+	let mut input = Input::new_from_chars("".chars());
 	let not_followed_parser = not_followed_by(&parser);
 	let output = not_followed_parser(&mut input);
 	assert_eq!(output, Err(Error::EndOfInput));
@@ -13,7 +13,7 @@ fn empty() {
 #[test]
 fn followed() {
 	let parser = is('h');
-	let mut input = Input::new("h".chars());
+	let mut input = Input::new_from_chars("h".chars());
 	let not_followed_parser = not_followed_by(&parser);
 	let output = not_followed_parser(&mut input);
 	assert_eq!(output, Err(Error::UnexpectedToken(Position::new(1, 1))));
@@ -22,7 +22,7 @@ fn followed() {
 #[test]
 fn not_followed() {
 	let parser = is('h');
-	let mut input = Input::new("jello".chars());
+	let mut input = Input::new_from_chars("jello".chars());
 	let not_followed_parser = not_followed_by(&parser);
 	let output = not_followed_parser(&mut input);
 	assert_eq!(output, Ok(()));
@@ -32,13 +32,13 @@ fn not_followed() {
 fn look_ahead_followed() {
 	// Inspiration: https://github.com/haskell/parsec/issues/8
 	let parser = is('h');
-	let mut input = Input::new("hello".chars());
+	let mut input = Input::new_from_chars("hello".chars());
 	let lookahead_parser = look_ahead(&parser);
 	// Just ensure that it succeeds to prove a point.
 	let output = lookahead_parser(&mut input);
 	assert_eq!(output, Ok('h'));
 	// Actually test.
-	let mut input = Input::new("hello".chars());
+	let mut input = Input::new_from_chars("hello".chars());
 	let not_followed_parser = not_followed_by(&lookahead_parser);
 	let output = not_followed_parser(&mut input);
 	assert_eq!(output, Err(Error::UnexpectedToken(Position::new(1, 1))));
