+++ src/tests/option.rs
@@ -5,7 +5,7 @@ use input::position::Position;
 fn success_first() {
 	let parser1 = is('h');
 	let parser2 = is('j');
-	let mut input = Input::new("h".chars());
+	let mut input = Input::new_from_chars("h".chars());
 	let parse_option = option(&parser1, &parser2);
 	let output = parse_option(&mut input).unwrap();
 	assert_eq!(output, 'h');
@@ -16,7 +16,7 @@ fn success_first() {
 fn success_second() {
 	let parser1 = is('h');
 	let parser2 = is('j');
-	let mut input = Input::new("j".chars());
+	let mut input = Input::new_from_chars("j".chars());
 	let parse_option = option(&parser1, &parser2);
 	let output = parse_option(&mut input).unwrap();
 	assert_eq!(output, 'j');
@@ -27,7 +27,7 @@ fn success_second() {
 fn fail_not_consuming() {
 	let parser1 = is('h');
 	let parser2 = is('j');
-	let mut input = Input::new("kello".chars());
+	let mut input = Input::new_from_chars("kello".chars());
 	let parse_option = option(&parser1, &parser2);
 	assert_eq!(
 		parse_option(&mut input),
@@ -40,7 +40,7 @@ fn fail_not_consuming() {
 fn fail_consuming() {
 	let parser1 = is('h');
 	let parser2 = is('j');
-	let mut input = Input::new("hello".chars());
+	let mut input = Input::new_from_chars("hello".chars());
 	let consuming_parser = |input: &mut Input<_>| {
 		parser1(input)?;
 		parser2(input)
