+++ src/tests/satisfy.rs
@@ -11,11 +11,11 @@ fn digits() {
 		}
 	});
 	// Digits.
-	let mut input = Input::new("1".chars());
+	let mut input = Input::new_from_chars("1".chars());
 	assert_eq!(parser(&mut input), Ok('1'));
 	assert!(end_of_input()(&mut input).is_ok());
 	// Words fails and does not consume.
-	let mut input = Input::new("hello".chars());
+	let mut input = Input::new_from_chars("hello".chars());
 	assert_eq!(
 		parser(&mut input),
 		Err(Error::UnexpectedToken(Position::new(1, 1)))
